requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
python-dotenv>=1.0.0
httpx>=0.24.1
pymupdf>=1.22.5
pydantic>=2.0.0
python-multipart>=0.0.6
groq>=0.4.0
cohere>=4.32
aiofiles>=23.1.0
loguru>=0.7.0
tenacity>=8.2.0
requests>=2.31.0
python-telegram-bot>=20.6.0
